feat: add delete hooks for new repositories

+++ api/src/nlp/repositories/nlp-metric.repository.ts
@@ -41,6 +41,20 @@ export class NlpMetricRepository extends BaseRepository<
     super(eventEmitter, model, NlpMetric, NLP_METRIC_POPULATE, NlpMetricFull);
   }
 
+  private async updateOrDeleteModel(theId: string): Promise<void> {
+    const updatedModel = await this.nlpModelRepository.updateOne(
+      { experiments: theId },
+      {
+        $pull: { experiments: theId },
+        $inc: { version: -1 },
+      },
+    );
+
+    if (updatedModel?.version <= 0 && updatedModel?.experiments.length === 0) {
+      await this.nlpModelRepository.deleteOne({ _id: updatedModel.id });
+    }
+  }
+
   /**
    * Deletes NLP experiment associated with the provided criteria before deleting the metrics themselves.
    *
@@ -59,9 +73,44 @@ export class NlpMetricRepository extends BaseRepository<
   ): Promise<void> {
     {
       if (criteria._id || criteria.name) {
-        await this.nlpMetricValueRepository.deleteMany({
-          metric: criteria.name,
-        });
+        // Find associated metric values
+        const metricValues = (
+          await this.nlpMetricValueRepository.find({ metric: criteria.name })
+        ).map((doc) => ({ metric: doc.metric, value: doc.value }));
+
+        if (metricValues.length > 0) {
+          // Find experiments associated with the metrics
+          const experimentsIds = (
+            await this.nlpExperimentRepository.find({
+              metrics: {
+                $in: metricValues.map((val) => {
+                  val.metric, val.value;
+                }),
+              },
+            })
+          ).map((doc) => doc.id);
+          if (experimentsIds.length > 0) {
+            // Update or delete model associated with the experiments
+            await Promise.all(
+              experimentsIds.map(async (theId) => {
+                await this.updateOrDeleteModel(theId);
+              }),
+            );
+            // Remove the processed experiments
+
+            await this.nlpExperimentRepository.deleteMany({
+              _id: { $in: experimentsIds },
+            });
+          }
+          // Remove the processed metric values
+          await this.nlpMetricValueRepository.deleteMany({
+            metric: {
+              $in: metricValues.map((val) => {
+                val.metric, val.value;
+              }),
+            },
+          });
+        }
       } else {
         throw new Error(
           'Attempted to delete NLP metrics using unknown criteria',
